Turning plain ROOT macro for likelihood fit into a class: LHFitter
